Reducing RoutingWeight size from 24 bytes to 16 bytes.

+++ routing/index_graph_starter.cpp
@@ -150,7 +150,7 @@ bool IndexGraphStarter::CheckLength(RouteWeight const & weight)
 {
   // We allow 1 pass-through/non-pass-through zone changes per ending located in
   // non-pass-through zone to allow user to leave this zone.
-  int32_t const numPassThroughChangesAllowed =
+  int8_t const numPassThroughChangesAllowed =
       (StartPassThroughAllowed() ? 0 : 1) + (FinishPassThroughAllowed() ? 0 : 1);
 
   return weight.GetNumPassThroughChanges() <= numPassThroughChangesAllowed &&
